@ Extract integration tests to standalone glassfish-tests module
- Create glassfish-tests module with shared test infrastructure
- Unify duplicated test code from glassfish-managed and glassfish-remote
- Add Maven profiles: glassfish-managed (default), glassfish-remote
- Add glassfish8 profile to override glassfish.version to 8.0.4
- Refactor remote tests to use @ArquillianResource URL instead of hardcoded localhost
- Add docker-compose.yaml with GF 7 and GF 8 services
- Add 4 CI jobs: gf7-managed, gf7-remote, gf8-managed, gf8-remote
- Remove test code from glassfish-managed and glassfish-remote modules

@
