+++ .github/workflows/build.yml
@@ -1,12 +1,4 @@
-# This workflow will build a Java project with Maven, and cache/restore any dependencies to improve the workflow execution time
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-java-with-maven
-
-# This workflow uses actions that are not certified by GitHub.
-# They are provided by a third-party and are governed by
-# separate terms of service, privacy policy, and support
-# documentation.
-
-name: Build
+name: Build and Test
 
 on:
   push:
@@ -25,12 +17,94 @@ jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v3
-    - name: Set up JDK 21
-      uses: actions/setup-java@v3
-      with:
-        java-version: 21
-        distribution: "temurin"
-        cache: "maven"
-    - name: Build with Maven
-      run: ./mvnw -B package --file pom.xml
+      - uses: actions/checkout@v4
+      - name: Set up JDK 21
+        uses: actions/setup-java@v4
+        with:
+          java-version: 21
+          distribution: "temurin"
+          cache: "maven"
+      - name: Build with Maven (skip tests)
+        run: ./mvnw -B install -DskipTests --file pom.xml
+
+  gf7-managed:
+    needs: build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up JDK 21
+        uses: actions/setup-java@v4
+        with:
+          java-version: 21
+          distribution: "temurin"
+          cache: "maven"
+      - name: Run GF 7 Managed Tests
+        run: ./mvnw -B verify -pl glassfish-tests --file pom.xml
+
+  gf7-remote:
+    needs: build
+    runs-on: ubuntu-latest
+    services:
+      glassfish:
+        image: omnifish/glassfish:7.0.14
+        ports:
+          - 4848:4848
+          - 8080:8080
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up JDK 21
+        uses: actions/setup-java@v4
+        with:
+          java-version: 21
+          distribution: "temurin"
+          cache: "maven"
+      - name: Wait for GlassFish
+        run: |
+          for i in $(seq 1 30); do
+            curl -sf http://localhost:4848 && break
+            echo "Waiting for GlassFish admin listener..."
+            sleep 10
+          done
+      - name: Run GF 7 Remote Tests
+        run: ./mvnw -B verify -pl glassfish-tests -Pglassfish-remote --file pom.xml
+
+  gf8-managed:
+    needs: build
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up JDK 21
+        uses: actions/setup-java@v4
+        with:
+          java-version: 21
+          distribution: "temurin"
+          cache: "maven"
+      - name: Run GF 8 Managed Tests
+        run: ./mvnw -B verify -pl glassfish-tests -Pglassfish-managed,glassfish8 --file pom.xml
+
+  gf8-remote:
+    needs: build
+    runs-on: ubuntu-latest
+    services:
+      glassfish:
+        image: omnifish/glassfish:8.0.4
+        ports:
+          - 4848:4848
+          - 8080:8080
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up JDK 21
+        uses: actions/setup-java@v4
+        with:
+          java-version: 21
+          distribution: "temurin"
+          cache: "maven"
+      - name: Wait for GlassFish
+        run: |
+          for i in $(seq 1 30); do
+            curl -sf http://localhost:4848 && break
+            echo "Waiting for GlassFish admin listener..."
+            sleep 10
+          done
+      - name: Run GF 8 Remote Tests
+        run: ./mvnw -B verify -pl glassfish-tests -Pglassfish-remote,glassfish8 --file pom.xml
