--- glassfish-managed/src/test/resources/arquillian.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0"?>
-<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns="http://jboss.org/schema/arquillian"
-    xsi:schemaLocation="http://jboss.org/schema/arquillian http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
-
-  <!-- Uncomment to have test archives exported to the file system for inspection -->
-  <engine>
-      <property name="deploymentExportPath">target/</property>
-  </engine>
-
-  <container qualifier="glassfish" default="true">
-    <configuration>
-      <property name="glassFishHome">target/${glassfish.home}</property>
-      <property name="adminHost">localhost</property>
-      <property name="adminPort">4848</property>
-      <property name="debug">true</property>
-      <property name="enableDerby">${enableDerby:true}</property>
-      <property name="allowConnectingToRunningServer">false</property>
-      <property name="outputToConsole">true</property>
-      <!-- true enables verbose logging of the jaxrs requests -->
-      <property name="debugRequests">true</property>
-    </configuration>
-  </container>
-
-  <container qualifier="glassfish-auth">
-    <configuration>
-      <property name="glassFishHome">target/${glassfish.home}</property>
-      <property name="adminHost">localhost</property>
-      <property name="adminPort">4848</property>
-      <property name="adminUser">admin</property>
-      <property name="adminPassword">admin</property>
-    </configuration>
-  </container>
-
-</arquillian>
