--- glassfish-remote/src/test/java/org/jboss/arquillian/container/glassfish/remote/CDIJarTestCase.java
@@ -1,49 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2011, Red Hat Middleware LLC, and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jboss.arquillian.container.glassfish.remote;
-
-import jakarta.inject.Inject;
-import org.jboss.arquillian.container.test.api.Deployment;
-import org.jboss.arquillian.junit5.ArquillianExtension;
-import org.jboss.shrinkwrap.api.ShrinkWrap;
-import org.jboss.shrinkwrap.api.asset.EmptyAsset;
-import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-@ExtendWith(ArquillianExtension.class)
-public class CDIJarTestCase {
-
-    @Inject
-    private SimpleBean foo;
-
-    @Deployment
-    public static JavaArchive deploy() {
-        return ShrinkWrap
-            .create(JavaArchive.class, "foo.jar")
-            .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml")
-            .addClasses(SimpleBean.class, CDIJarTestCase.class);
-    }
-
-    @Test
-    public void test() {
-        assertNotNull(foo);
-    }
-}
-
