--- glassfish-remote/src/test/java/org/jboss/arquillian/container/glassfish/remote/GreeterServlet.java
@@ -1,48 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2011, Red Hat Middleware LLC, and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @author <a href="http://community.jboss.org/people/LightGuard">Jason Porter</a>
- */
-package org.jboss.arquillian.container.glassfish.remote;
-
-import jakarta.ejb.EJB;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-/**
- * Simple servlet for testing deployment.
- *
- * @author <a href="http://community.jboss.org/people/aslak">Aslak Knutsen</a>
- * @author <a href="http://community.jboss.org/people/LightGuard">Jason Porter</a>
- */
-@WebServlet(urlPatterns = "/Greeter")
-public class GreeterServlet extends HttpServlet {
-    private static final long serialVersionUID = 8249673615048070666L;
-
-    @EJB
-    private Greeter greeter;
-
-    @Override
-    protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append(this.greeter.greet());
-    }
-}
