--- glassfish-remote/src/test/resources/arquillian.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0"?>
-<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns="http://jboss.org/schema/arquillian"
-    xsi:schemaLocation="http://jboss.org/schema/arquillian http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
-
-  <!-- Uncomment to have test archives exported to the file system for inspection -->
-  <!--
-  <engine>
-      <property name="deploymentExportPath">target/</property>
-  </engine>
-  -->
-
-  <container qualifier="glassfish" default="true">
-    <configuration>
-      <property name="adminHost">localhost</property>
-      <property name="adminPort">4848</property>
-      <!-- true enables verbose logging of the jaxrs requests -->
-      <property name="debugRequests">false</property>
-    </configuration>
-  </container>
-
-  <container qualifier="glassfish-auth">
-    <configuration>
-      <property name="adminHost">localhost</property>
-      <property name="adminPort">4848</property>
-      <property name="adminUser">admin</property>
-      <property name="adminPassword">admin</property>
-    </configuration>
-  </container>
-
-</arquillian>
