+++ glassfish-tests/docker-compose.yaml
@@ -1,0 +1,14 @@
+services:
+  glassfish7:
+    image: omnifish/glassfish:7.0.14
+    profiles: ["gf7"]
+    ports:
+      - "8080:8080"
+      - "4848:4848"
+
+  glassfish8:
+    image: omnifish/glassfish:8.0.4
+    profiles: ["gf8"]
+    ports:
+      - "8081:8080"
+      - "4849:4848"
