+++ glassfish-tests/src/test/resources/application.xml
