+++ glassfish-tests/src/test/resources/arq-glassfish-managed/arquillian.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0"?>
+<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xmlns="http://jboss.org/schema/arquillian"
+    xsi:schemaLocation="http://jboss.org/schema/arquillian http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
+
+  <defaultProtocol type="Servlet 5.0"/>
+
+  <engine>
+      <property name="deploymentExportPath">target/</property>
+  </engine>
+
+  <container qualifier="glassfish-managed" default="true">
+    <configuration>
+      <property name="glassFishHome">target/${glassfish.home}</property>
+      <property name="adminHost">localhost</property>
+      <property name="adminPort">4848</property>
+      <property name="debug">true</property>
+      <property name="enableDerby">${enableDerby:true}</property>
+      <property name="allowConnectingToRunningServer">false</property>
+      <property name="outputToConsole">true</property>
+      <property name="debugRequests">true</property>
+    </configuration>
+  </container>
+
+</arquillian>
