+++ glassfish-tests/src/test/resources/arq-glassfish-remote/arquillian.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0"?>
+<arquillian xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xmlns="http://jboss.org/schema/arquillian"
+    xsi:schemaLocation="http://jboss.org/schema/arquillian http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
+
+  <defaultProtocol type="Servlet 5.0"/>
+
+  <engine>
+      <property name="deploymentExportPath">target/</property>
+  </engine>
+
+  <container qualifier="glassfish-remote" default="true">
+    <configuration>
+      <property name="adminHost">${glassfish.adminHost:localhost}</property>
+      <property name="adminPort">${glassfish.adminPort:4848}</property>
+      <property name="adminUser">${glassfish.adminUser:admin}</property>
+      <property name="adminPassword">${glassfish.adminPassword:}</property>
+      <property name="debugRequests">false</property>
+    </configuration>
+  </container>
+
+</arquillian>
