+++ pom.xml
@@ -55,6 +55,7 @@
     <maven-javadoc-plugin.version>3.12.0</maven-javadoc-plugin.version>
     <maven-source-plugin.version>3.3.1</maven-source-plugin.version>
     <maven-gpg-plugin.version>3.2.7</maven-gpg-plugin.version>
+    <build-helper-maven-plugin.version>3.6.0</build-helper-maven-plugin.version>
 
     <!-- Arquillian & tests -->
     <arquillian.version>1.10.0.Final</arquillian.version>
@@ -74,6 +75,7 @@
     <subproject>glassfish-common</subproject>
     <subproject>glassfish-managed</subproject>
     <subproject>glassfish-remote</subproject>
+    <subproject>glassfish-tests</subproject>
   </subprojects>
 
   <dependencyManagement>
@@ -149,6 +151,16 @@
           <version>${maven-compiler-plugin.version}</version>
         </plugin>
       
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-failsafe-plugin</artifactId>
+          <version>${maven-surefire-plugin.version}</version>
+        </plugin>
+        <plugin>
+          <groupId>org.codehaus.mojo</groupId>
+          <artifactId>build-helper-maven-plugin</artifactId>
+          <version>${build-helper-maven-plugin.version}</version>
+        </plugin>
         <plugin>
           <groupId>org.apache.maven.plugins</groupId>
           <artifactId>maven-shade-plugin</artifactId>
